+++ js/components/home/home.jsx
@@ -1,0 +1,3 @@
+export function Home() {
+  return <h1 className="text-center text-blue-700">Hello, world</h1>;
+}