+++ tailwind.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  mode: 'jit',
+  content: ['./js/**/*.{js,jsx}', './static/dist/**/*.{js,css}', './app/templates/base.html'],
+  darkMode: 'class',
+}